environmentspawner optimized
the environmentspawner spawns in all props (currently only crystals and cargoboxes) in the level. before this commit, it spawned objects at beginplay. the problem was that they were always loaded (even if they were "spatially loaded"). after this commit, only the objects' locations are set at beginplay, and are spawned when in proximity with the boat.